ref(connection): Improve replication client semantics and structure
