+++ crates/etl/src/replication/client/mod.rs
@@ -1,0 +1,12 @@
+mod child;
+mod query;
+mod raw;
+mod transaction;
+mod types;
+mod utils;
+
+pub use child::ChildPgReplicationClient;
+pub use raw::PgReplicationClient;
+pub use transaction::{PgChildReplicationTransaction, PgReplicationTransaction};
+pub use types::{CreateSlotResult, CtidPartition, GetSlotResult, SlotState};
+pub(crate) use types::{GetOrCreateSlotResult, PostgresConnectionUpdate};
